# move directory existence check for code statistics task inside of the…