… task

otherwise, if a gem appends to the constant, we can end up with an error.
Such an example is the `view_component` gem:  
https://github.com/ViewComponent/view_component/blob/main/lib/view_component/rails/tasks/view_component.rake#L3

It is a dependency of one of my dependencies, so I don't have the `app/components` directory inside my app, and when running `rails stats`, it fails due to that:

```
rails stats                  
rails aborted!
Errno::ENOENT: No such file or directory @ dir_initialize - app/components

Tasks: TOP => stats
(See full trace by running task with --trace)
```
### Checklist

Before submitting the PR make sure the following are checked:

* [x] This Pull Request is related to one change. Changes that are unrelated should be opened in separate PRs.
* [x] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [x] Tests are added or updated if you fix a bug or add a feature.
* [ ] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.