@p8 2023-02-08
The whole code block for "assigning the Constant and checking for the directory existence"  got moved out of the rake task in: 272f80dc16505fd52b7abfabdbba14192b608ba7 to make the constant available for other libraries.

The checks to see if all these directories exist run for every rake task (`bin/rails db:migrate`, `bin/rails test`).
This seems wasteful, so moving it into the task seems like a good idea.

But I'm not sure how the existing code could throw an error if the directories don't exist:
```ruby
ViewComponent::Base.view_component_path # => "app/components"
File.directory?("app/components") # => false
```