@gregmolnar 2023-02-08
@p8 it shouldn't throw an error in my opinion but it does when trying to loop over the directories here: https://github.com/rails/rails/blob/main/railties/lib/rails/code_statistics.rb#L46