@gregmolnar 2023-02-08
And the reason I think it shouldn't error is because as in my case, a dependency of a dependency can add directories to the constant that the actual app is not intending to have. But not having that directory, I can't get the stats, as it fails to find the directory.