@p8 2023-02-08
Some quick performance metrics:
Before this change:

     ( for i in {1..25}; do; bin/rails -h; done; )  18.48s user 6.77s system 97% cpu 25.791 total

After this change:

     ( for i in {1..25}; do; bin/rails -h; done; )  18.11s user 6.67s system 97% cpu 25.320 total