@p8 2023-02-08
> @p8 it shouldn't throw an error in my opinion but it does when trying to loop over the directories here: https://github.com/rails/rails/blob/main/railties/lib/rails/code_statistics.rb#L46

Ah, ok that makes sense. So currently we:
* define the directories and filter non existing and assign that to `STATS_DIRECTORIES`
* ViewComponent adds `app/components` to `STATS_DIRECTORIES` but that directory isn't filtered.
* An error is thrown trying to use this directory