@gregmolnar 2023-02-08
Yes, that's roughly what happens. Rails does the directory existence check before View Components appends to the constant, so the directories added by it are not checked for existence. 
I didn't even realize the performance implications of running the check outside of the task, nice catch!