@p8 2023-02-09
Might be good to add a test for this, similar to: https://github.com/rails/rails/blob/main/railties/test/commands/dev_test.rb